+++ argser/__init__.py
@@ -1,9 +1,10 @@
 from argser.consts import FALSE_VALUES, TRUE_VALUES
 from argser.display import make_table, stringify
 from argser.fields import Arg, Opt
-from argser.parse_func import call
+from argser.parse_func import call, SubCommands
 from argser.parser import parse_args, sub_command
 
 parse = parse_args
 Argument = Arg
 Option = Opt
+Subs = SubCommands